Fixed Backspace and newline characters
Backspace now will not crash the program in the event of pressing
backspace at the top of the file.

Enter now also makes the cursor go to the correct line instead of
the line below it.

I also edited comments to make them more accurate to what they do
now.